+++ src/index.ts
@@ -1,16 +1,2 @@
-export {
-  installConsoleLogger,
-  type ConsoleLoggerHandle,
-  type ConsoleLoggerOptions,
-} from "./logger/consoleLogger";
-
-export { installNetworkLogger, type NetworkLoggerHandle, type NetworkLoggerOptions } from "./logger/networkLogger";
-
-export { LogStorage, type LogStorageOptions } from "./storage/logStorage";
-
-export { createPanel, type PanelHandle, type PanelOptions } from "./ui/panel";
-
-export { init, initDevInspector, type DevInspectorHandle, type DevInspectorInitOptions } from "./init";
-
-export type { ConsoleLogLevel, LogEntry, LogSource } from "./utils/types";
+export { initDevInspector } from "./init";